piwigo-videojs: Add date validation
The plugin was blindly accepting the results of strtotime which, when
the date is zero, returned a negative value. This caused updates of the
database to fail.

This adds a helper function, check_date, that will return the date in
YYYY-MM-DD HH:MM:SS format (code borrowed from the core PWG metadata
handling) or it will return Null if the timestamp is zero.